refactor: Change File Directory By Android Architecture
