+++ app/src/androidTest/java/campus/tech/kakao/map/MapErrorActivityTest.kt
@@ -5,7 +5,7 @@ import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.rules.ActivityScenarioRule
-import campus.tech.kakao.map.view.MapErrorActivity
+import campus.tech.kakao.map.ui.map_error.MapErrorActivity
 import org.junit.Rule
 import org.junit.Test
 import org.junit.Assert.*
