+++ app/src/main/java/campus/tech/kakao/map/ui/map/MapViewModel.kt
@@ -1,20 +1,20 @@
-package campus.tech.kakao.map.viewmodel
+package campus.tech.kakao.map.ui.map
 
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import campus.tech.kakao.map.model.DataStoreRepository
+import campus.tech.kakao.map.data.last_location.LastLocationRepository
 import kotlinx.coroutines.launch
 import java.io.IOException
 
 class MapViewModel(
-    private val dataStoreRepository: DataStoreRepository
+    private val lastLocationRepository: LastLocationRepository
 ) : ViewModel() {
 
     fun saveLastLocation(latitude: Double, longitude: Double) {
         viewModelScope.launch {
             try {
-                dataStoreRepository.saveLocation(latitude, longitude)
+                lastLocationRepository.saveLocation(latitude, longitude)
             } catch (ioException: IOException) {
                 Log.e("MapViewModel", "Failed to save location", ioException)
             } catch (exception: Exception) {
@@ -25,7 +25,7 @@ class MapViewModel(
 
     fun loadLastLocation(callback: (Double, Double, Boolean) -> Unit) {
         viewModelScope.launch {
-            dataStoreRepository.loadLocation().collect {
+            lastLocationRepository.loadLocation().collect {
                 it?.let {
                     callback(it.latitude, it.longitude, false)
                 }
