+++ app/src/main/java/campus/tech/kakao/map/ui/map/MapViewModelFactory.kt
@@ -1,15 +1,15 @@
-package campus.tech.kakao.map.viewmodel
+package campus.tech.kakao.map.ui.map
 
 import android.content.Context
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import campus.tech.kakao.map.model.DataStoreRepository
+import campus.tech.kakao.map.data.last_location.LastLocationRepository
 
 @Suppress("UNCHECKED_CAST")
 class MapViewModelFactory(private val context: Context) : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         if (modelClass.isAssignableFrom(MapViewModel::class.java)) {
-            val repository = DataStoreRepository(context)
+            val repository = LastLocationRepository(context)
             return MapViewModel(repository) as T
         }
         throw IllegalArgumentException("Unknown ViewModel class")
