+++ app/src/main/java/campus/tech/kakao/map/ui/search/SearchLocationViewModel.kt
@@ -1,12 +1,12 @@
-package campus.tech.kakao.map.viewmodel
+package campus.tech.kakao.map.ui.search
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import campus.tech.kakao.map.model.HistoryRepository
-import campus.tech.kakao.map.model.Location
-import campus.tech.kakao.map.model.SearchLocationRepository
+import campus.tech.kakao.map.data.history.HistoryRepository
+import campus.tech.kakao.map.data.local_search.Location
+import campus.tech.kakao.map.data.local_search.SearchLocationRepository
 import kotlinx.coroutines.launch
 
 class SearchLocationViewModel(
