+++ app/src/main/res/layout/activity_map.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.MapActivity">
+    tools:context=".ui.map.MapActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/main"
