+++ app/src/test/java/campus/tech/kakao/map/MapViewModelTest.kt
@@ -1,8 +1,8 @@
 package campus.tech.kakao.map
 
 import android.util.Log
-import campus.tech.kakao.map.model.DataStoreRepository
-import campus.tech.kakao.map.viewmodel.MapViewModel
+import campus.tech.kakao.map.data.last_location.LastLocationRepository
+import campus.tech.kakao.map.ui.map.MapViewModel
 import com.kakao.vectormap.LatLng
 import io.mockk.Runs
 import io.mockk.coEvery
@@ -25,7 +25,7 @@ import java.io.IOException
 
 @ExperimentalCoroutinesApi
 class MapViewModelTest {
-    private val mockRepository = mockk<DataStoreRepository>()
+    private val mockRepository = mockk<LastLocationRepository>()
     private val viewModel = MapViewModel(mockRepository)
 
     @Before
