+++ app/src/test/java/campus/tech/kakao/map/SearchLocationRepositoryTest.kt
@@ -3,13 +3,12 @@ package campus.tech.kakao.map
 import android.content.ContentValues
 import android.database.Cursor
 import android.database.sqlite.SQLiteDatabase
-import campus.tech.kakao.map.model.HistoryDbHelper
-import campus.tech.kakao.map.model.LocalSearchDocument
-import campus.tech.kakao.map.model.LocalSearchMeta
-import campus.tech.kakao.map.model.LocalSearchResponse
-import campus.tech.kakao.map.model.LocalSearchSameName
-import campus.tech.kakao.map.model.LocalSearchService
-import campus.tech.kakao.map.model.SearchLocationRepository
+import campus.tech.kakao.map.data.local_search.LocalSearchDocument
+import campus.tech.kakao.map.data.local_search.LocalSearchMeta
+import campus.tech.kakao.map.data.local_search.LocalSearchResponse
+import campus.tech.kakao.map.data.local_search.LocalSearchSameName
+import campus.tech.kakao.map.data.local_search.LocalSearchService
+import campus.tech.kakao.map.data.local_search.SearchLocationRepository
 import io.mockk.Runs
 import io.mockk.coEvery
 import io.mockk.every
