+++ app/src/test/java/campus/tech/kakao/map/SearchLocationViewModelTest.kt
@@ -2,9 +2,9 @@ package campus.tech.kakao.map
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Observer
-import campus.tech.kakao.map.model.Location
-import campus.tech.kakao.map.model.SearchLocationRepository
-import campus.tech.kakao.map.viewmodel.SearchLocationViewModel
+import campus.tech.kakao.map.data.local_search.Location
+import campus.tech.kakao.map.data.local_search.SearchLocationRepository
+import campus.tech.kakao.map.ui.search.SearchLocationViewModel
 import io.mockk.Runs
 import io.mockk.coEvery
 import io.mockk.coVerify
